pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-avp"
version = "0.1.0"
description = "OpenClaw credential provider using Agent Vault Protocol"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    {name = "Omar Abdelmalek", email = "abdelmalek.omar1@gmail.com"}
]
keywords = ["openclaw", "avp", "vault", "credentials", "ai-agents", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "avp-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.urls]
Homepage = "https://github.com/avp-protocol/openclaw-avp"
Repository = "https://github.com/avp-protocol/openclaw-avp"

[tool.setuptools.packages.find]
where = ["src"]
